requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.0
numpy==1.25.2
lxml==4.9.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
joblib==1.3.2
scipy==1.11.1
threadpoolctl==3.2.0
soupsieve==2.4.1
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
urllib3==2.0.4
google-generativeai==0.3.2